Removed GraphQL support (sob!)
