+++ db/blip_sync_messages.go
@@ -40,7 +40,6 @@ const (
 	MessagePutRev       = "putRev"       // Connected Client API
 	MessageUnsubChanges = "unsubChanges" // Connected Client API
 	MessageFunction     = "function"     // Connected Client API
-	MessageGraphQL      = "graphql"      // Connected Client API
 )
 
 // Message properties
@@ -116,10 +115,6 @@ const (
 	// query (Connected Client)
 	QueryName   = "name"
 	QuerySource = "src"
-
-	// graphQL (Connected Client)
-	GraphQLQuery         = "query"
-	GraphQLOperationName = "operationName"
 
 	// Sync Gateway specific properties (used for testing)
 	SGShowHandler = "sgShowHandler" // Used to request a response with sgHandler
